% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_helper_functions.R
\name{assign_behavior}
\alias{assign_behavior}
\title{Assign behavior estimates to observations}
\usage{
assign_behavior(dat.list, theta.estim.long, behav.names)
}
\arguments{
\item{dat.list}{A list of data associated with each animal ID where names of
list elements are the ID names.}

\item{theta.estim.long}{A data frame in long format where each observation
(\emph{time1}) of each time segment (\emph{tseg}) of each animal ID
(\emph{id}) has separate rows for behavior proportion estimates per state.
Columns for behavior and proportion estimates should be labeled
\emph{behavior} and \emph{prop}, respectively. Date (in POSIXct format)
should also be included as a column labeled \emph{date}.}

\item{behav.names}{character. A vector of names to label each state (in
order).}
}
\value{
A data frame of all animal IDs where columns (with names from
  \code{behav.names}) include proportions of each behavioral state per
  observation, as well as a column that stores the dominant behavior within a
  given time segment for which the observation belongs (\code{behav}).
}
\description{
Assign behavior estimates to observations
}
\examples{

}
