% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CumSumInv}
\alias{CumSumInv}
\title{Internal function that calculates the inverted cumsum}
\usage{
CumSumInv(ntsegm, nmaxclust, z)
}
\arguments{
\item{ntsegm}{An integer.}

\item{nmaxclust}{An integer.}

\item{z}{An integer matrix.}
}
\description{
Internal function that calculates the inverted cumsum
}
