% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_helper_functions.R
\name{df_to_list}
\alias{df_to_list}
\title{Convert data frame to a list by animal ID}
\usage{
df_to_list(dat, ind)
}
\arguments{
\item{dat}{A data frame containing the data for each animal ID.}

\item{ind}{character. The name of the column storing the animal IDs.}
}
\value{
A list where each element stores the data for a separate animal ID.
}
\description{
Converts an object of class \code{data.frame} to a list where each element is
a separate animal ID. This function prepares the data for further analysis
and when mapping other functions onto the data for separate animal IDs.
}
\examples{
#simulate data
step<- rgamma(1000, c(1, 2.5, 10), c(1, 1, 1))
angle<- runif(1000, -pi, pi)
date<- seq(c(ISOdate(2020, 6, 17, tz = "UTC")), by = "hour", length.out = 1000)
date<- date + lubridate::seconds(runif(length(date), -15, 15))  #introduce noise
dt<- as.numeric(diff(date)) * 60  #convert time difference to seconds
dt<- c(dt, NA)
id<- rep(1:10, each = 100)

#create data frame
dat<- data.frame(id, date, dt, step, angle)
dat<- round_track_time(dat = dat, id = "id", dt = "dt", date = "date", int = 3600,
                       tol = 15, time.zone = "UTC")

#run function
dat.list<- df_to_list(dat = dat, ind = "id")


}
