% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{discrete.move.var}
\alias{discrete.move.var}
\title{Discretize movement variables}
\usage{
discrete.move.var(dat, lims, varIn, varOut)
}
\arguments{
\item{dat}{A data frame that contains the variable(s) of interest to convert
from continuous to discrete values.}

\item{lims}{A list of the bin limits for each variable. Each element of the
list should be a vector of real numbers.}

\item{varIn}{A vector of names for the continuous variable stored as columns
within \code{dat}.}

\item{varOut}{A vector of names for the storage of the discrete variables
returned by the function.}
}
\value{
A data frame with new columns of discretized variables as labelled by
  \code{varOut}.
}
\description{
Convert movement variables from continuous to discrete values for analysis by
\code{segment.behavior}.
}
\examples{


}
