% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_helper_functions.R
\name{discrete_move_var}
\alias{discrete_move_var}
\title{Discretize movement variables}
\usage{
discrete_move_var(dat, lims, varIn, varOut)
}
\arguments{
\item{dat}{A data frame that contains the variable(s) of interest to convert
from continuous to discrete values.}

\item{lims}{A list of the bin limits for each variable. Each element of the
list should be a vector of real numbers.}

\item{varIn}{A vector of names for the continuous variable stored as columns
within \code{dat}.}

\item{varOut}{A vector of names for the storage of the discrete variables
returned by the function.}
}
\value{
A data frame with new columns of discretized variables as labeled by
  \code{varOut}.
}
\description{
Convert movement variables from continuous to discrete values for analysis by
\code{\link{segment_behavior}}.
}
\examples{
#simulate data
step<- rgamma(1000, c(1, 2.5, 10), c(1, 1, 1))
angle<- runif(1000, -pi, pi)
id<- rep(1:10, each = 100)

#create data frame
dat<- data.frame(id, step, angle)

#define limits for each bin
dist.lims<- c(quantile(step, c(0, 0.25, 0.5, 0.75, 0.95)), max(step))
angle.lims<- c(-pi, -3*pi/4, -pi/2, -pi/4, 0, pi/4, pi/2, 3*pi/4, pi)

#run function
dat1<- discrete_move_var(dat = dat, lims = list(dist.lims, angle.lims),
                         varIn = c("step", "angle"),
                         varOut = c("SL","TA"))


}
