% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_helper_functions.R
\name{expand_behavior}
\alias{expand_behavior}
\title{Expand behavior estimates from time segments to observations}
\usage{
expand_behavior(dat, theta.estim, nobs)
}
\arguments{
\item{dat}{A data frame of the animal ID, time segment labels, and all other
data per observation. Animal ID and time segment columns must be labeled
\emph{id} and \emph{tseg}, respectively.}

\item{theta.estim}{A data frame containing the animal ID, time segment, and
proportions of each behavioral state as separate columns. Animal ID and
time segment columns must be labeled \emph{id} and \emph{tseg},
respectively.}

\item{nobs}{A data frame containing the animal ID, time segment, and number
of observations recorded within a given time segment. Animal ID, time
segment, and sample size columns must be labeled \emph{id}, \emph{tseg},
and \emph{n}, respectively.}
}
\value{
A new data frame that expands behavior proportions for each
  observation within all time segments, including the columns labeled
  \emph{time1} and \emph{date} from the original \code{dat} data frame.
}
\description{
Expand behavior estimates from time segments to observations
}
\examples{

}
