% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_helper_functions.R
\name{get_MAP_internal}
\alias{get_MAP_internal}
\title{Internal function to find the maximum a posteriori (MAP) estimate of the MCMC
chain}
\usage{
get_MAP_internal(dat, nburn)
}
\arguments{
\item{dat}{numeric. A vector of log marginal likelihood values for a given
animal ID.}

\item{nburn}{numeric. The size of the burn-in phase after which the MAP
estimate will be identified.}
}
\value{
A numeric value indicating the iteration after the burn-in phase that
  holds the MAP estimate.
}
\description{
Internal function to be used by a wrapper.
}
