% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_helper_functions.R
\name{plot_heatmap_behav}
\alias{plot_heatmap_behav}
\title{Internal function for plotting breakpoints over a heatmap of the discretized
movement variables}
\usage{
plot_heatmap_behav(data, nbins, brkpts, title, legend)
}
\arguments{
\item{data}{A data frame for a single animal ID that contains only columns
for the ID and each of the movement variables that were analyzed by
\code{\link{segment_behavior}}. The ID column must be first.}

\item{nbins}{numeric. A vector of the number of bins used to discretize each
movement variable. These must be in the same order as the columns within
\code{data}.}

\item{brkpts}{A data frame that contains the breakpoints associated with each
animal ID. This data frame is returned by \code{\link{get_breakpts}}.}

\item{title}{logical. If \code{TRUE}, includes the animal ID as the title of
the plot.}

\item{legend}{logical. If \code{TRUE}, shows the legend for the heatmap.}
}
\value{
A heatmap for each movement variable showing the bins assigned to
  each observation over time that is overlayed by the extracted breakpoints.
}
\description{
An internal function for plotting the results of the segmentation model.
}
