% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_helper_functions.R
\name{prep_data}
\alias{prep_data}
\title{Calculate step lengths and turning angles from coordinates}
\usage{
prep_data(dat, coord.names, id)
}
\arguments{
\item{dat}{A data frame that contains a column for animal IDs and the columns
associated with the x and y coordinates. For easier interpretation of the
model results, it is recommended that coordinates be stored after UTM
projection (meters) as opposed to unprojected in decimal degrees (map
units).}

\item{coord.names}{character. A vector of the column names under which the
coordinates are stored. The name for the x coordinate should be listed
first and the name for the y coordinate second.}

\item{id}{character. The name of the column storing the animal IDs.}
}
\value{
A data frame where all original data are returned and new columns are
  added for step length (\code{step}) and turning angle (\code{angle}).
}
\description{
Calculates step lengths and turning angles based on coordinates for each
animal ID. Provides a self-contained method to calculate these variables
without needing to rely on other R packages (e.g., \code{adehabitatLT}).
However, functions from other packages can also be used to perform this step
in data preparation.
}
\examples{

}
