% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultinom1}
\alias{rmultinom1}
\title{Internal function that samples z's from a categorical distribution}
\usage{
rmultinom1(prob, randu)
}
\arguments{
\item{prob}{A numeric matrix.}

\item{randu}{A numeric vector.}
}
\description{
Internal function that samples z's from a categorical distribution
}
