% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultinom2}
\alias{rmultinom2}
\title{Internal function that samples z's from a multinomial distribution}
\usage{
rmultinom2(prob, n, randu, nmaxclust)
}
\arguments{
\item{prob}{A numeric vector.}

\item{n}{An integer.}

\item{randu}{A numeric vector.}

\item{nmaxclust}{An integer.}
}
\description{
Internal function that samples z's from a multinomial distribution
}
