% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmod_gibbs_functions.R
\name{sample.gamma.mixmod}
\alias{sample.gamma.mixmod}
\title{Internal function to sample the gamma hyperparameter}
\usage{
sample.gamma.mixmod(v, ngroup, gamma.possib)
}
\arguments{
\item{v}{numeric. A vector of proportions for each of the possible clusters.}

\item{ngroup}{numeric. The total number of possible clusters.}

\item{gamma.possib}{numeric. A vector of possible values that gamma can take
ranging between 0.1 and 1.}
}
\value{
A single numeric value for gamma that falls within
  \code{gamma.possib} for calculation of the log-likelihood.
}
\description{
Internal function to sample the gamma hyperparameter
}
