% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmod_gibbs_functions.R
\name{sample.v.mixmod}
\alias{sample.v.mixmod}
\title{Internal function to sample parameter for truncated stick-breaking prior}
\usage{
sample.v.mixmod(z, gamma1, nmaxclust)
}
\arguments{
\item{z}{A vector of latent cluster estimates provided by
\code{\link{sample.z.mixmod}}.}

\item{gamma1}{numeric. Hyperparameter for the truncated stick-breaking prior.}

\item{nmaxclust}{numeric. A single number indicating the maximum number of
clusters to test.}
}
\value{
A list with estimates for \emph{v} and \emph{theta} for each of the possible states.
}
\description{
This function samples the latent \emph{v} parameter within the Gibbs sampler.
Not for calling directly by users.
}
