% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SampleZAgg}
\alias{SampleZAgg}
\title{Internal function that samples z1 aggregate}
\usage{
SampleZAgg(ntsegm, b1, y1, nmaxclust, lphi1, ltheta, zeroes)
}
\arguments{
\item{ntsegm}{An integer.}

\item{b1}{An integer.}

\item{y1}{An integer matrix.}

\item{nmaxclust}{An integer.}

\item{lphi1}{A numeric matrix.}

\item{ltheta}{A numeric matrix.}

\item{zeroes}{A numeric vector.}
}
\description{
Internal function that samples z1 aggregate
}
