% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StoreZ}
\alias{StoreZ}
\title{This function helps store z from all iterations after burn in}
\usage{
StoreZ(z, store_z, nobs)
}
\arguments{
\item{z}{An integer vector.}

\item{store_z}{An integer matrix.}

\item{nobs}{An integer.}
}
\description{
This function helps store z from all iterations after burn in
}
