% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summarize1}
\alias{summarize1}
\title{Internal function that summarizes bin distributions of track segments}
\usage{
summarize1(VecVals, Breakpts, nobs, nbins, nbreak)
}
\arguments{
\item{VecVals}{A vector of bin values.}

\item{Breakpts}{A vector if breakpoints.}

\item{nobs}{The number of observations.}

\item{nbins}{The number of bins for a given data stream.}

\item{nbreak}{The number of estimated breakpoints.}
}
\description{
Internal function that summarizes bin distributions of track segments
}
