% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_helper_functions.R
\name{summarize_tsegs}
\alias{summarize_tsegs}
\title{Summarize observations within bins per time segment}
\usage{
summarize_tsegs(dat, nbins)
}
\arguments{
\item{dat}{A data frame of the animal ID, time segment number, and the
discretized data for each movement variable. Animal ID and time segment
must be the first two columns of this data frame.}

\item{nbins}{numeric. A vector of the number of bins used to discretize each
movement variable. These must be in the same order as the columns within
\code{dat}.}
}
\value{
A new data frame that contains the animal ID, time segment number,
  and the counts per bin for each movement variable. The names for each of
  these bins are labeled according to the order in which the variables were
  provided to \code{summarize_tsegs}.
}
\description{
Pepares the data that has already been segmented for clustering by Latent
Dirichlet Allocation. This function summarizes the counts observed per
movement variable bin within each time segment per animal ID.
}
\examples{

}
