% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SummarizeDat}
\alias{SummarizeDat}
\title{Internal function that generates nmat matrix to help with multinomial draws}
\usage{
SummarizeDat(z, dat, ncateg, nbehav, nobs)
}
\arguments{
\item{z}{An integer vector.}

\item{dat}{An integer vector.}

\item{ncateg}{An integer.}

\item{nbehav}{An integer.}

\item{nobs}{An integer.}
}
\description{
Internal function that generates nmat matrix to help with multinomial draws
}
