% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\docType{data}
\name{tracks}
\alias{tracks}
\title{Simulated set of three tracks.}
\format{
A data frame with 15003 rows and 4 variables: \describe{
\item{id}{ID for each simulated track}
\item{date}{date, recorded as datetime}
\item{x}{x coordinate of tracks}
\item{y}{y coordinate of tracks} }
}
\usage{
tracks
}
\description{
A dataset containing the IDs as well as x and y coordinates for three tracks
of 5001 observations each (15,003 in total).
}
\keyword{datasets}
