% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_helper_functions.R
\name{filter_time}
\alias{filter_time}
\title{Filter observations for time interval of interest}
\usage{
filter_time(dat.list, int)
}
\arguments{
\item{dat.list}{A list of data associated with each animal ID where names of
list elements are the ID names.}

\item{int}{numeric. The time interval of interest.}
}
\value{
A list where observations for each animal ID (element) has been
  filtered for \code{int}. Two columns (\code{obs} and \code{time1}) are
  added for each list element (ID), which store the original observation
  number before filtering and the new observation number after filtering,
  respectively.
}
\description{
Selects observations that belong to the time interval of interest and removes
all others. This function also removes entire IDs from the dataset when there
is one or fewer observations at this time interval. This function works
closely with \code{\link{round_track_time}} to only retain observations
sampled at a regular time interval, which is important for analyzing step
lengths and turning angles. Column storing the time intervals must be labeled
\code{dt}.
}
\examples{
#simulate data
step<- rgamma(1000, c(1, 2.5, 10), c(1, 1, 1))
angle<- runif(1000, -pi, pi)
date<- seq(c(ISOdate(2020, 6, 17, tz = "UTC")), by = "hour", length.out = 1000)
date<- date + lubridate::seconds(runif(length(date), -15, 15))  #introduce noise
dt<- as.numeric(diff(date)) * 60  #convert time difference to seconds
dt<- c(dt, NA)
id<- rep(1:10, each = 100)

#create data frame
dat<- data.frame(id, date, dt, step, angle)
dat<- round_track_time(dat = dat, id = "id", int = 3600, tol = 15, time.zone = "UTC")

#create list
dat.list<- df_to_list(dat = dat, ind = "id")

#run function
dat.list.filt<- filter_time(dat.list = dat.list, int = 3600)

}
