% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_helper_functions.R
\name{get_MAP}
\alias{get_MAP}
\title{Find the maximum a posteriori (MAP) estimate of the MCMC chain}
\usage{
get_MAP(dat, nburn)
}
\arguments{
\item{dat}{A data frame where each row holds the log marginal likelihood
values at each iteration of the MCMC chain.}

\item{nburn}{numeric. The size of the burn-in phase after which the MAP
estimate will be identified.}
}
\value{
A numeric vector of iterations at which the MAP estimate was found
  for each animal ID.
}
\description{
Identify the MCMC iteration that holds the MAP estimate. This will be used to
inform \code{\link{get_breakpts}} as to which breakpoints should be retained
on which to assign track segments to the observations of each animal ID.
}
\examples{
#simulate data
ngibbs<- 1000
y<- (-1000 * 501:1500)/(-500 + 501:1500) + rnorm(ngibbs, 0, 0.1)
dat<- matrix(c(1, y), 1, 1001)
dat<- data.frame(dat)
names(dat)[1]<- "id"

#run function
MAP.est<- get_MAP(dat = dat, nburn = ngibbs/2)

}
